# accelerate core ruby requires