--- a/config/patches/ruby/ruby-fast-load_26.patch
+++ b/config/patches/ruby/ruby-fast-load_26.patch
@@ -1,0 +1,16 @@
+diff --git a/load.c b/load.c
+index 576464fb68..ae89f63820 100644
+--- a/load.c
++++ b/load.c
+@@ -908,6 +908,11 @@ search_required(VALUE fname, volatile VALUE *path, int safe_level, feature_func
+        if (loading) *path = rb_filesystem_str_new_cstr(loading);
+        return 'r';
+     }
++    else if ((ft = rb_feature_p(ftptr, 0, FALSE, FALSE, &loading)) == 's') {
++       if (loading) *path = rb_filesystem_str_new_cstr(loading);
++       return 's';
++    }
++
+     tmp = fname;
+     type = rb_find_file_ext_safe(&tmp, loadable_ext, safe_level);
+     switch (type) {