see https://bugs.ruby-lang.org/issues/15856