@lamont-granquist 2019-11-02
This is a 15% speedup in `chef --version`, not as impactful to chef-client/knife with all the require_relative and requires guards:

```
with patch:

% time /opt/chef/bin/knife --version
Chef Infra Client: 15.4.58
/opt/chef/bin/knife --version  0.60s user 0.11s system 97% cpu 0.734 total
% time /opt/chef/bin/chef-client --version
Chef Infra Client: 15.4.58
/opt/chef/bin/chef-client --version  1.61s user 0.22s system 98% cpu 1.867 total

% time /opt/chefdk/bin/chef --version
ChefDK version: 4.4.26
Chef Infra Client version: 15.3.14
Chef InSpec version: 4.16.0
Test Kitchen version: 2.3.3
Foodcritic version: 16.1.1
Cookstyle version: 5.6.2
/opt/chefdk/bin/chef --version  6.04s user 2.63s system 98% cpu 8.764 total


without patch:

i12:26 crash chef/omnibus % time /opt/chef/bin/knife --version
Chef Infra Client: 15.4.58
/opt/chef/bin/knife --version  0.70s user 0.11s system 97% cpu 0.837 total
git:chef:master↓✗ rvm:ruby-2.5.3
12:28 crash chef/omnibus % time /opt/chef/bin/chef-client --version
Chef Infra Client: 15.4.58
/opt/chef/bin/chef-client --version  1.65s user 0.26s system 98% cpu 1.935 total

i3:29 trimix chef-dk/omnibus % time /opt/chefdk/bin/chef --version
ChefDK version: 4.4.26
Chef Infra Client version: 15.3.14
Chef InSpec version: 4.16.0
Test Kitchen version: 2.3.3
Foodcritic version: 16.1.1
Cookstyle version: 5.6.2
/opt/chefdk/bin/chef --version  7.02s user 3.15s system 99% cpu 10.257 total
```

This breaks a ruby spec that doesn't seem to be useful at all where if you require a c-extension first before requiring the ruby extension that the ruby extension will not load.

There are some gems, including internal ones which do have both .so and .rb files like openssl.rb/openssl.so -- but in those cases they load the ruby file first and then load the .so, so this patch does not break that behavior.

The spec dates back over 10 years to the inception of the ruby/spec project with no documentation as to what the use case / design consideration actually was.