@lamont-granquist 2019-11-02
and validated my thinking against @zenspider and he couldn't come up with a problem with this either.